*: add SavedPartitions filtering by index range and label glob
Add --save-partlabel <glob> and --save-partindex <range>, where <range>
can be a single value or a single-ended or double-ended range.  The
option names aren't great, but they fit in the 80-character limit.
Multiples of each option can be specified, and/or multiple filters can be
specified within a single option argument by separating them with commas.
If no filters are specified, no partitions are saved.

Add corresponding kargs which, unlike the command-line args, cannot be
repeated.  Multiple comma-separated filters can be specified within a
karg value.

Drop the --wipedisk option, since saving nothing is now the default.

Upon restore, try to reuse the original partition number, but if we
can't, renumber the partition to one more than the highest number used
so far.  For simplicity, never backfill entries earlier in the partition
table, even if the corresponding slot is unused.